+++ packages/admin/src/index.js
@@ -1,4 +1,0 @@
-export { default as AdminApp } from './AdminApp.jsx'
-export { default as Tabs } from './Tabs.jsx'
-// Consumers must also import the stylesheet once:
-//   import '@speakeasy/admin/admin.css'